+++ Changelog.md
@@ -1,6 +1,7 @@
 TenshiLib 1.9.6
 ================
 - Fix some things with animation gui and suggestion box
+- Add selection list widget
 
 TenshiLib 1.9.5
 ================